Ce 006 school

+++ src/school/school.controller.ts
@@ -19,11 +19,17 @@ import { UpdateSchoolDTO } from './dto/updateSchool.dto';
 import { AuthGuard } from '@nestjs/passport';
 import { IRequest } from 'src/auth/interface/request.interface';
 import { Role } from '@prisma/client';
+import { CreatePermissionSchoolDTO } from 'src/permission/dto/createPermissionSchool.dto';
+import { UpdatePermissionSchoolDTO } from 'src/permission/dto/upatePermissionSchool.dto';
+import { PermissionService } from 'src/permission/permission.service';
 
 @ApiTags('School')
 @Controller('school')
 export class SchoolController {
-  constructor(private readonly schoolService: SchoolService) {}
+  constructor(
+    private readonly schoolService: SchoolService,
+    private readonly permissionService: PermissionService,
+  ) {}
 
   @UseGuards(AuthGuard('jwt'))
   @ApiOperation({ summary: 'Get All School (Admin)' })
@@ -69,7 +75,7 @@ export class SchoolController {
 
   @UseGuards(AuthGuard('jwt'))
   @ApiOperation({ summary: 'Create School (Admin)' })
-  @Post('create')
+  @Post()
   async createSchool(
     @Request() req: IRequest,
     @Body() createSchoolDTO: CreateSchoolDTO,
@@ -85,6 +91,69 @@ export class SchoolController {
     return {
       message: 'Successfully Create School',
       data: school,
+    };
+  }
+
+  @UseGuards(AuthGuard('jwt'))
+  @ApiOperation({ summary: 'Create Permission School (Admin)' })
+  @Post('permission')
+  async createPermissionSchool(
+    @Request() req: IRequest,
+    @Body()
+    createPermissionSchoolDTO: CreatePermissionSchoolDTO,
+  ) {
+    if (req.user.role !== Role.ADMIN) {
+      throw new HttpException(
+        'Do Not Have Permission(Admin)',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    const invalidSchool = await this.schoolService.getSchoolById(
+      createPermissionSchoolDTO.schoolId,
+    );
+    if (!invalidSchool) {
+      throw new HttpException('School Not Found', HttpStatus.NOT_FOUND);
+    }
+    const school = await this.permissionService.createPermisssionSchoolById(
+      createPermissionSchoolDTO,
+    );
+    return {
+      message: 'Successfully Create Permission School',
+      data: school,
+    };
+  }
+
+  @UseGuards(AuthGuard('jwt'))
+  @ApiOperation({ summary: 'Edit Permission School By Id (Admin)' })
+  @Put('permission/edit/:id')
+  async updatePermissionSchoolById(
+    @Request() req: IRequest,
+    @Param('id', ParseUUIDPipe) id: string,
+    @Body() updatePermissionSchoolDTO: UpdatePermissionSchoolDTO,
+  ) {
+    if (req.user.role !== Role.ADMIN) {
+      throw new HttpException(
+        'Do Not Have Permission(Admin)',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    const permission = await this.permissionService.updatePermissionSchoolById(
+      updatePermissionSchoolDTO,
+      id,
+    );
+
+    if (!permission) {
+      throw new HttpException(
+        'Permission School Not Found',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    return {
+      message: 'Update Permission School Successfully',
+      data: permission,
     };
   }
 
@@ -127,6 +196,13 @@ export class SchoolController {
         HttpStatus.FORBIDDEN,
       );
     }
+    const invalidSchool = await this.schoolService.getSchoolById(id);
+    if (!invalidSchool) {
+      throw new HttpException('School Not Found', HttpStatus.NOT_FOUND);
+    }
+    await this.permissionService.deletePermissionSchoolByPermisssionId(
+      invalidSchool?.permission?.permissionId,
+    );
 
     const school = await this.schoolService.deleteSchoolById(id);
     if (!school) {
